[kernel] ata driver use uint32_t for block count instead of uint8_t

+++ src/device/ata_driver.cpp
@@ -117,7 +117,7 @@ void ata_driver::irq_handle(uint64_t irq_handle_num)
     }
 }
 lock_type ata_lock = {0};
-void ata_driver::read(uint32_t where, uint8_t count, uint8_t *buffer)
+void ata_driver::read(uint32_t where, uint32_t count, uint8_t *buffer)
 {
     lock(&ata_lock);
     waiting_for_irq = 1;
@@ -131,7 +131,7 @@ void ata_driver::read(uint32_t where, uint8_t count, uint8_t *buffer)
 
     ata_write(true, ATA_reg_command_status, 0x20); // call the read command
 
-    int new_count = count;
+    uint32_t new_count = count;
     uint32_t off = 0;
     uint64_t wait_time = 0;
     while (new_count-- > 0)
